// frontend/src/components/CreateExamForm/enums/Subjects.ts
// Klausuren-Fächer

export enum Subjects {
  Englisch = "Englisch",
  Deutsch = "Deutsch",
  Ethik = "Ethik",
  PoWi = "PoWi",
  LF1 = "LF1",
  LF2 = "LF2",
  LF3 = "LF3",
  LF4 = "LF4",
  LF5 = "LF5",
  LF6 = "LF6",
  LF7 = "LF7",
  LF8 = "LF8",
  LF9 = "LF9",
  LF10 = "LF10",
  LF11 = "LF11",
  LF12 = "LF12",
}